Trying to switch around onchangehandler
