+++ src/Home.jsx
@@ -2,9 +2,22 @@ import React, { Component } from "react";
 import DisplayProductData from "./components/DisplayProductData";
 import { Icon } from 'semantic-ui-react'
 import DisplayCategoriesList from './components/DisplayCategoriesList'
+import { getData } from "./modules/productData";
 
 class Home extends Component {
+  state = {
+		productData: [],
+	};
 
+  componentDidMount() {
+		this.getProductData();
+	}
+
+	async getProductData() {
+		let result = await getData();
+		this.setState({ productData: result });
+  }
+  
   render() {
 
   return (
@@ -12,7 +25,7 @@ class Home extends Component {
     <div id="container" as='h1'>Revisarik<Icon name='food' size='large' />
       </div>
       <div id="menu">
-        <DisplayProductData />
+        <DisplayProductData productData={this.state.productData}/>
         <DisplayCategoriesList/>
      </div>
     </>
